backend/requirements.txt
Django==4.2.11
djangorestframework==3.15.2
django-cors-headers==4.4.0

psycopg2-binary==2.9.9
redis==5.0.7
rq==1.16.2
rq-scheduler==0.13.1

pandas==2.2.2
openpyxl==3.1.5
pdfplumber==0.11.4
fpdf2==2.7.9
